+++ R/ggplot_hbar.R
@@ -235,8 +235,8 @@ ggplot_hbar <- function(data,
 #' @param y_title Y axis title string. Defaults to [Y title].
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. Not applicable where mobile equals TRUE.
 #' @param col_labels Adjust the  x scale labels through a vector.
-#' @param col_labels_ncol The number of columns in the legend. 
-#' @param col_labels_nrow The number of rows in the legend.
+#' @param col_legend_ncol The number of columns in the legend. 
+#' @param col_legend_nrow The number of rows in the legend.
 #' @param col_na TRUE or FALSE of whether to show NA values of the colour variable. Defaults to TRUE.
 #' @param col_rev TRUE or FALSE of whether bar fill order from left to right is reversed from default. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL.
@@ -295,8 +295,8 @@ ggplot_hbar_col <-
            y_title = "[Y title]",
            y_title_wrap = 50,
            col_labels = NULL,
-           col_labels_ncol = NULL,
-           col_labels_nrow = NULL,
+           col_legend_ncol = NULL,
+           col_legend_nrow = NULL,
            col_na = TRUE,
            col_rev = FALSE,
            col_title = "",
@@ -471,13 +471,13 @@ ggplot_hbar_col <-
           caption = stringr::str_wrap(caption, caption_wrap)
         ) +
         guides(fill = guide_legend(
-          ncol = col_labels_ncol, nrow = col_labels_nrow, 
+          ncol = col_legend_ncol, nrow = col_legend_nrow, 
           byrow = TRUE,
           reverse = TRUE,
           title = stringr::str_wrap(col_title, col_title_wrap)
         ), 
         col = guide_legend(
-          ncol = col_labels_ncol, nrow = col_labels_nrow, 
+          ncol = col_legend_ncol, nrow = col_legend_nrow, 
           byrow = TRUE,
           reverse = TRUE,
           title = stringr::str_wrap(col_title, col_title_wrap)
@@ -731,8 +731,8 @@ ggplot_hbar_facet <-
 #' @param y_title Y axis title string. Defaults to [Y title].
 #' @param y_title_wrap Number of characters to wrap the y title to. Defaults to 50. 
 #' @param col_labels Adjust the  x scale labels through a vector.
-#' @param col_labels_ncol The number of columns in the legend. 
-#' @param col_labels_nrow The number of rows in the legend.
+#' @param col_legend_ncol The number of columns in the legend. 
+#' @param col_legend_nrow The number of rows in the legend.
 #' @param col_na TRUE or FALSE of whether to show NA values of the colour variable. Defaults to TRUE.
 #' @param col_rev TRUE or FALSE of whether bar fill order from left to right is reversed from default. Defaults to FALSE.
 #' @param col_title Colour title string for the legend. Defaults to NULL.
@@ -793,8 +793,8 @@ ggplot_hbar_col_facet <-
            y_title = "[Y title]",
            y_title_wrap = 50,
            col_labels = NULL,
-           col_labels_ncol = NULL,
-           col_labels_nrow = NULL,
+           col_legend_ncol = NULL,
+           col_legend_nrow = NULL,
            col_na = TRUE,
            col_rev = FALSE,
            col_title = "",
@@ -959,13 +959,13 @@ ggplot_hbar_col_facet <-
       ) +
       facet_wrap(vars(!!facet_var), scales = facet_scales, ncol = facet_ncol, nrow = facet_nrow) +
       guides(fill = guide_legend(
-        ncol = col_labels_ncol, nrow = col_labels_nrow, 
+        ncol = col_legend_ncol, nrow = col_legend_nrow, 
         byrow = TRUE,
         reverse = TRUE,
         title = stringr::str_wrap(col_title, col_title_wrap)
       ), 
       col = guide_legend(
-        ncol = col_labels_ncol, nrow = col_labels_nrow, 
+        ncol = col_legend_ncol, nrow = col_legend_nrow, 
         byrow = TRUE,
         reverse = TRUE,
         title = stringr::str_wrap(col_title, col_title_wrap)