+++ NEWS.md
@@ -1,3 +1,9 @@
+# simplevis 3.1.0
+
+* Breaking change: Changed col_labels_nrow and col_labels_ncol to col_legend_ncol and col_legend_nrow.
+* Breaking change: Added plotly_col_legend function, and removed plotly_legend_rev and plotly_legend_order.
+* Breaking change: leaflet popup_var has been removed, and replaced with a popup_vars_vctr argument.
+
 # simplevis 3.0.0
 
 * Bumped version for CRAN release.